+++ eatr/src/main/java/nayanda/droid/eatr/base/BaseHttpRequestWithBody.java
@@ -17,6 +17,14 @@ public abstract class BaseHttpRequestWithBody<T extends BaseHttpRequestWithBody>
 
     protected String body;
 
+    protected BaseHttpRequestWithBody(String method) {
+        super(method);
+    }
+
+    String getBody() {
+        return body;
+    }
+
     @NonNull
     @Override
     @SuppressWarnings("unchecked")