+++ eatr/src/main/java/nayanda/droid/eatr/builder/HttpGet.java
@@ -3,6 +3,8 @@ package nayanda.droid.eatr.builder;
 import android.support.annotation.NonNull;
 
 import nayanda.droid.eatr.base.BaseHttpRequest;
+import nayanda.droid.eatr.base.HttpRequestExecutorImpl;
+import nayanda.droid.eatr.base.HttpRequestRestExecutorImpl;
 import nayanda.droid.eatr.digester.Digester;
 import nayanda.droid.eatr.digester.Finisher;
 import nayanda.droid.eatr.digester.ProgressDigester;
@@ -13,47 +15,51 @@ import nayanda.droid.eatr.digester.RestResponse;
  * Created by nayanda on 08/02/18.
  */
 
-class HttpGet extends BaseHttpRequest<HttpGet> {
+public class HttpGet extends BaseHttpRequest<HttpGet> {
+
+    HttpGet() {
+        super("GET");
+    }
 
     @Override
-    public void asyncExecute(@NonNull ProgressDigester<Response> responseProgressDigester) {
-        asyncExecutor("GET", null, responseProgressDigester);
+    public void asyncExecute(@NonNull final ProgressDigester<Response> responseProgressDigester) {
+        asyncExecutor(method, null, new HttpRequestExecutorImpl(this, responseProgressDigester));
     }
 
     @Override
     public <O> void asyncExecute(@NonNull ProgressDigester<RestResponse<O>> restResponseProgressDigester, @NonNull Class<O> withModelClass) {
-        asyncExecutor(withModelClass, "GET", null, restResponseProgressDigester);
+        asyncExecutor(method, null, new HttpRequestRestExecutorImpl<>(this, withModelClass, restResponseProgressDigester));
     }
 
     @Override
     public void asyncExecute(@NonNull final Digester<Response> responseDigester) {
-        asyncExecutor("GET", null, responseDigester);
+        asyncExecutor(method, null, new HttpRequestExecutorImpl(this, responseDigester));
     }
 
     @Override
     public <O> void asyncExecute(@NonNull final Digester<RestResponse<O>> restResponseDigester, @NonNull final Class<O> withModelClass) {
-        asyncExecutor(withModelClass, "GET", null, restResponseDigester);
+        asyncExecutor(method, null, new HttpRequestRestExecutorImpl<>(this, withModelClass, restResponseDigester));
     }
 
     @Override
     public void asyncExecute(@NonNull final Finisher<Response> responseFinisher) {
-        asyncExecutor("GET", null, responseFinisher);
+        asyncExecutor(method, null, new HttpRequestExecutorImpl(this, responseFinisher));
     }
 
     @Override
     public <O> void asyncExecute(@NonNull final Finisher<RestResponse<O>> restResponseFinisher, @NonNull final Class<O> withModelClass) {
-        asyncExecutor(withModelClass, "GET", null, restResponseFinisher);
+        asyncExecutor(method, null, new HttpRequestRestExecutorImpl<>(this, withModelClass, restResponseFinisher));
     }
 
     @NonNull
     @Override
     public Response execute() {
-        return executor("GET");
+        return executor(method);
     }
 
     @NonNull
     @Override
     public <O> RestResponse<O> execute(@NonNull Class<O> withModelClass) {
-        return executor(withModelClass, "GET");
+        return executor(withModelClass, method);
     }
 }