+++ eatr/src/main/java/nayanda/droid/eatr/digester/Consumer.java
@@ -1,0 +1,11 @@
+package nayanda.droid.eatr.digester;
+
+/**
+ * Created by nayanda on 15/02/18.
+ */
+
+public interface Consumer<T> {
+
+    void onConsume(T param);
+
+}