+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ## [Released]
+### 0.1.0 -2018-02-15
+#### Added
+- HttpRequestExecutor for separated consumer
+#### Edited
+- Simplified executor on BaseHttpRequest class
 
 ### 0.0.10 -2018-02-13
 #### Added